makeFields -> makeClassy
This change seems to clean up the lens-based code quite nicely, and only
at the cost of manually defining a few extra instances. Also removed the
logger since we're not using it (likely better to just use MonadThrow in
some capacity, we don't actually need/want to log computed values).