Add simple exception handling
It didn't make sense to continue execution if there was a parse error.
Since points are more than likely being sent programmatically, if there
is one parse error, then there are probably going to be a lot more, so
it's best to just kill it right away and let the user know there's a
problem.